frontend: Add initial scaffolding
The Elm application is basically what you get from running `elm init`
and then copying the SPA example[^1]. The content has been slightly
adjusted, to flex with centering a text (no pun intended).

The backend directly references the JS bundle to include it in its
binary. The following command needs to be executed before building
the server (from the Nix dev shell):

```sh
elm make src/Main.elm --output static/main.js
```

Elm can also transpile to html directly, but someone decided that a
different font is an absolute requirement... To adjust the html head
tag's content, a custom html file is necessary.

Maud is used for the html template to avoid additional files that need
to be included into the build. This probably increases compile times,
but since the big sins syn and proc-macro2 are already in the dependency
graph anyways, might as well just pretend that everything is fine.

[^1]: https://guide.elm-lang.org/webapps/navigation#example