server: Read bundle at runtime from given path
Instead of including the JS file at compile time, the server now
accepts a path to the bundle directory as CLI argument.

If no such argument is given, the server simply disables the UI
frontend. This allows working on the API or the Unix client without
even knowing that a frontend exists (ignorance is bliss).

The server now also no longer uses maud for the `index.html`. The
reason for this is twofold:
 - using an actual file is much easier to integrate with `ServeDir`
 - having the file available on the filesystem allows the frontend
   to be hosted with any static file server (which for example
   enables hot reloading of the frontend during development)

Since this change already touched like 90% of the existing code in
the web module, I took the opportunity to replace warp with axum.
Previous discussions concluded that there was no specific reason for
using warp and that we should switch to axum instead to benefit from
its larger ecosystem.