server: Expand derivation build state model
Discussions in #32 showed that we will likely need to mark some
derivation builds as blocked, if they depend on another derivation
that has previously been interrupted and not yet retried. Added a
new state to accomodate for that.

Previous discussions also revealed that the state type must be
representable as a single database column, so they all use new types
instead of struct types now - which better captures this intention.

We also discussed to remove `DrvBuildResult` in favour of a single
boolean field. A single boolean wrapped in a new type however felt
not expressive enough (i.e. it is not immediately clear what exactly
the boolean stands for). I therefore decided to keep it as a separate
enum for now.