server: Fix insert logic for derivation build
The `Drv` table was pretty much replaced by `DrvBuildModel`, so the
insert function had to be adjusted accordingly.

To implement this cleanly, I had to make sure all types of fields
that correspond to a database column implement `sqlx::Type`. For
the Git types this meant wrapping the actual types in a Newtype,
because of the orphan rule.

server: Add encode/decode impl for Git types