server: Autoselect build attempt for new metadata
This is absolutely necessary. Otherwise we *will* end up with wrong
entries in the database.

I faced the problem that I wanted to reuse the `DrvBuildMetadata`
struct for passing the initial values to the insert function, but
some of the fields were no longer meant to be set manually. Making
`build_attempt` an `Option` is also not a good solution, because when
selecting, it will never be `None`, so we would either have to deal
with an error case that we can never reach or have unwraps all over
the place. Instead I came up with the `ForInsert` Newtype, which can
only be constructed from inside the db module and only the db module
can access its inner value. This prevents other modules to access
potentially "unitialized" (filled with dummy values) fields.

All the trait implementations are just there to make the test code
work (i.e. make it possible to clone and compare stuff).