server: Add insert logic for DrvBuildEvent table
This required encoding the state into some value the database can
understand. Using an integer seemed logical, as a text would have
been a waste of space. I may have had some fun choosing the integer
values each state corresponds to.

Note that we can not use the `Type` derive macro with a `repr(i64)`
attribute, because of the nested enums, which can not be accurately
described like this.

It also turns out that I set the timestamp column to be of type text,
even though I set the default to `unixepoch` which returns an
integer. Because sqlx dynamically chooses how to decode a timestamp
based on the type, this caused problems. It is fixed now. Cheers to
the tests.