server: Expand possible build states
 - Added two new states `Buildable` and `TransitiveFailure`.
 - Renamed `Pending` state to `Queued`.
 - For every state, documented under what circumstances the
   derivation build leaves that state.

The `Buildable` state is an optimization that allows the server to
immediately make these derivations available to the builders on
restart. Otherwise, all derivations from the build queue would have
to go through the scheduler again on restart. During the discussion
of this change we also decided to rename the `Pending` state to
`Queued`, as we felt that the new name is more descriptive.

The `TransitiveFailure` state is required to allow us to mark
derivations whose build was never executed, because one of their
transitive dependencies completed with a failure. This has been a
requirement from the start, but I forgot about it when originally
adding this enum.