server: Stricter build state database conversion
The previous implementation required matching up the integer used to
represent a state accross two functions, which is error prone. The
new implementation instead uses an intermediate type that inlines all
possible states, which allows it to be `#[repr(u8)` and derive the
`sqlx::Type` trait directly. If a new state variant is added, the
compiler will enforce adding a two way conversion between the two
types, which is easier to review because the names match up.